backend/pyproject.toml
[project]
name = "ai-skin-cancer-detection-backend"
version = "0.1.0"
description = "AI Skin Cancer Detection System Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",

    "tensorflow>=2.15.0,<2.17.0",
    "opencv-python>=4.8.0,<4.10.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0,<2.0.0",

    "chromadb>=0.4.0",
    "onnxruntime==1.16.3",

    "torch==2.2.2",                 # IMPORTANT FIX
    "sentence-transformers>=2.2.0",

    "groq>=0.4.0",

    "langchain>=0.1.0",
    "langchain-groq>=0.0.1",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.0",

    "pypdf>=3.17.0",

    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]